Players can kill each other.
They are able to hit each other, play the appropriate animations & deal appropriate damage upon collision
