+++ FunGame/Assets/CharacterAssets/Valderheim/Animations/ValderheimAnims.controller
@@ -718,7 +718,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: 12892a9f8fcc8974bbfa0d03e59d9e29, type: 3}
+  m_Motion: {fileID: 7400000, guid: e54faca24ab60a24d9f0e8caff477ec3, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -746,7 +746,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: 7153116149701474f985eaf8ce9e4c2e, type: 2}
+  m_Motion: {fileID: 7400000, guid: e7153e460ee234f44b20cac2b9b292de, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -774,7 +774,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: f87cf3eb5e48ced478ebd3693f3d6c83, type: 3}
+  m_Motion: {fileID: 7400000, guid: 6dfa08ad7a800f740a7bf3c3cf88af4b, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -917,7 +917,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: 0ace4f570192ba944a12aa949e286628, type: 3}
+  m_Motion: {fileID: 7400000, guid: e9e63fbf74f0c1b4db5becfc0b53ed09, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -946,7 +946,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: 19a857adc04a0e94a829a92bfa74794e, type: 3}
+  m_Motion: {fileID: 7400000, guid: a01f2aa036e664a41a6f7da1a6b2df92, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -973,7 +973,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: c12dad602701cd246824322df00314da, type: 3}
+  m_Motion: {fileID: 7400000, guid: 80c8ead62793bda409deaefeafaff9b2, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter: 
@@ -1000,7 +1000,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: fab73e695bae6f441b5dec699c3da152, type: 3}
+  m_Motion: {fileID: 7400000, guid: f6cb96b7b865f404986d71a4a38b3e7c, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter:
