+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/ChargeRunEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6cb96b7b865f404986d71a4a38b3e7c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
