+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/HitblockoutEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7153e460ee234f44b20cac2b9b292de
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
