+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/ValderheimblockoutHitblockout.fbx.meta
@@ -150,23 +150,7 @@ ModelImporter:
       mirror: 0
       bodyMask: 01000000010000000100000001000000010000000100000001000000010000000100000001000000010000000100000001000000
       curves: []
-      events:
-      - time: 0.25474253
-        functionName: StartAttack()
-        data: 
-        objectReferenceParameter: {fileID: 11500000, guid: 1c04fbb76e9f97048a2f5ae62149bdc6,
-          type: 3}
-        floatParameter: 0
-        intParameter: 0
-        messageOptions: 0
-      - time: 0.8452088
-        functionName: EndAttack()
-        data: 
-        objectReferenceParameter: {fileID: 11500000, guid: 1c04fbb76e9f97048a2f5ae62149bdc6,
-          type: 3}
-        floatParameter: 0
-        intParameter: 0
-        messageOptions: 0
+      events: []
       transformMask: []
       maskType: 3
       maskSource: {instanceID: 0}
