+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/chargefinishblockEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e54faca24ab60a24d9f0e8caff477ec3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
