+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/chargeupblockEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e9e63fbf74f0c1b4db5becfc0b53ed09
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
