+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/frenzyEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6dfa08ad7a800f740a7bf3c3cf88af4b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
