+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/kickbackcomboblockEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 80c8ead62793bda409deaefeafaff9b2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
