+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/overheadblockEvent.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a01f2aa036e664a41a6f7da1a6b2df92
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
