+++ FunGame/Assets/CharacterAssets/Valderheim/Placeholderanims/runblockout.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b176f30045fb7c4cb22b24936a4ce39
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
