--- FunGame/Assets/Scripts/Players/Valderheim.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f832a189cdc4c3b4aaf6359d860657f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
