+++ FunGame/Assets/Scripts/Players/Valderheim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f72d6975e3c9cd2408c4dc82cadcfa0a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
